+++ src/features/home/index.ts
@@ -1,7 +1,4 @@
 export { TaskBasketButton } from "./ui/TaskBasketButton";
-export { TimeSlotList } from "./ui/TimeSlotList";
-export { TimeSlotGrid } from "./ui/TimeSlotGrid";
-export { TodoList } from "./ui/TodoList";
 export { TaskBasketAddSheet, TaskSplitSheet, TaskSplitSheetContent } from "./task-basket";
 export type { TaskSplitGroup, TaskSplitItem } from "./task-basket";
 export { useHomePlanner } from "./model/useHomePlanner";
