+++ src/widgets/planner-edit/ui/TimeSlotGrid.tsx
